ath11k-firmware: update IPQ8074 and QCN9074 to 2.9.0.1-02146
After a long time QCA has pushed an updated release of 2.9.0.1 firmware for IPQ8074 and QCN9074,
so lets update to 2.9.0.1-02146.

QCA has also moved the repository where they will be posting firmware to their CodeLinaro instance,
so we move to using that and it allows us to remove the manual download of QCN9074 board-2.bin.

Original commit :
Link: https://github.com/openwrt/openwrt/pull/16720
Signed-off-by: Robert Marko <robimarko@gmail.com>

Changes :
Remove support for IPQ6018,
not yet supported by OpenWrt-23.05.
Signed-off-by: Balsam CHIHI <balsam.chihi@moment.tech>